requirements.txt
numpy==1.24.4
pandas==2.0.3
Pillow==11.3.0
scikit_learn==1.3.2
torch==2.4.1
torchvision==0.19.1
tqdm==4.67.1
